requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
airflow-ai-sdk[openai]==0.1.7
apache-airflow-providers-amazon
pyyaml